Reorganization (cont.)
Moved presentation types to a separate library.

Reversed the direction of the core to/from persistence reference.

Using Autofac for managing services for the different libraries.

Re-coded the data context to be more of a data repository behind a set of interfaces. This was to completely hide the implementation details of the persistence layer.

Simplified validation code.